Create a new full chain cert when parts of the chain change
Problem:
When users were re-running clank, they were experiencing certificate mismatch
errors, because our ansible was not recreating the full chain. It wasn't
recreating it because the module performing the concatenation only
checked if the full chain file previously existed not whether it should be
recreated.

Solution:
Take a stroke from make, when ever the timestamps of the inputs are newer than
the output, regenerate the output.