Revert "Create a new full chain cert when parts of the chain change"
This reverts commit d455f35577af2a448976a6a0012c607ef646acc3.

After talking with @c-mart, it became apparent that this was an interesting
solution but overkill. I'm reverting it rather than nuking it, because it
is already published in several places and because it may be useful to refer to
at some point in the future where the complexity is warranted.